Logic
This sometimes gets stuck when testing with sudokus from sudoku.com on expert/evil, but honestly, I get stuck on said sudokus too and I am doubtful as to wether or not those puzzle are possible within reason; Wether or not anyone can have fun doing those by hand. The makers of Sudoku.com definitely do have a motive to make the impossible, as the site contains hints. First three for free, then, ads. It is ofcourse possible that I am just having a dumb moment.